Refactor Printful API to inherit from a base and update the YAML
