+++ app/interactors/api/printful/product.rb
@@ -2,12 +2,7 @@
 
 module Api
   module Printful
-    class Product
-      require "net/http"
-      require "net/https"
-
-      API_KEY = YAML.load_file("config/printful.yml")["api_key"]
-
+    class Product < Base
       attr_reader :id
 
       def initialize(id)
@@ -18,31 +13,10 @@ module Api
         new(id).perform
       end
 
-      def perform
-        request = build_products_request
-        response = http.start do |http|
-          http.request(request)
-        end
-        JSON.parse(response.body)
-      end
-
       private
 
-      def build_products_request
-        request = Net::HTTP::Get.new(products_uri.path)
-        request.add_field("Authorization", "Basic " + Base64.encode64(API_KEY).chomp)
-        request
-      end
-
-      def products_uri
-        @uri ||= URI.parse("https://api.printful.com/sync/products/#{id}")
-      end
-
-      def http
-        return @http if @http
-        @http = Net::HTTP.new(products_uri.host, products_uri.port)
-        @http.use_ssl = true
-        @http
+      def endpoint_path
+        "sync/products/#{id}"
       end
     end
   end
