+++ config/printful.yml.example
@@ -1,1 +1,2 @@
 api_key: abc123
+endpoint_url: https://api.printful.com/
